tf: Allow using large vpc (#278)
* tf: Allow configure vpc for ec2 tests

Need to pass down the variable all the way down from ec2 ->
basic_components > common ...

If we want to apply this to ecs, canary, soaking etc. we need to add
variable in those places as well

* tf: Move aoc vpc into common

It will be used by ecs, canaray etc. and we need to share it