Fix OTEL Operator version for E2E Operator EKS tests
We recently observed that main build was failing in ADOT Java repo, and failures started after May 20th, when OTEL Operator v0.100.0 was released, adding finalizer logic that we believe to be incompatible with our test framework. As a short term measure, we are fixing the OTEL Operator version to v0.99.0, the last known healthy version (as the test was passing on May 17th, when v0.99.0 was latest).